+++ src/pomdp-solve/pomdp-solve-options.c
@@ -337,33 +337,40 @@ POMDP_SOLVE_OPTS_showUsage( FILE* file, char* exec_name )
   /*******************************/
   fprintf( file, "General options:\n" );
 
-  PO_showUsageEnumType( file,
-                     POMDP_SOLVE_OPTS_ARG_FORCE_ROUNDING_STR,
-                     Boolean_Str );
+  fprintf( file, "\t%s <string>\n", POMDP_SOLVE_OPTS_ARG_POMDP_STR );
   fprintf( file, "\t%s <string>\n", POMDP_SOLVE_OPTS_ARG_STDOUT_STR );
   PO_showUsageEnumType( file,
                      POMDP_SOLVE_OPTS_ARG_SAVE_PENULTIMATE_STR,
                      Boolean_Str );
-  fprintf( file, "\t%s <string>\n", POMDP_SOLVE_OPTS_ARG_RAND_SEED_STR );
-  fprintf( file, "\t%s <string>\n", POMDP_SOLVE_OPTS_ARG_F_STR );
+  // fprintf( file, "\t%s <string>\n", POMDP_SOLVE_OPTS_ARG_F_STR );
   PO_showUsageEnumType( file,
                      POMDP_SOLVE_OPTS_ARG_STAT_SUMMARY_STR,
                      Boolean_Str );
+  PO_showUsageEnumType( file,
+                     POMDP_SOLVE_OPTS_ARG_VERBOSE_STR,
+                     POMDP_SOLVE_OPTS_Verbose_Str );
 
   /*******************************/
   /* Resource Limits parameters  */
   /*******************************/
+  /* Disabled in the R version
   fprintf( file, "Resource Limits options:\n" );
 
   fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_TIME_LIMIT_STR );
   fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_MEMORY_LIMIT_STR );
+  */
 
   /*******************************/
   /* Algorithm parameters  */
   /*******************************/
-  fprintf( file, "Algorithm options:\n" );
+  fprintf( file, "\nAlgorithm options:\n" );
 
-  fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_MCGS_PRUNE_FREQ_STR );
+  PO_showUsageEnumType( file,
+                     POMDP_SOLVE_OPTS_ARG_METHOD_STR,
+                     POMDP_SOLVE_OPTS_Method_Str );
+  PO_showUsageEnumType( file,
+                     POMDP_SOLVE_OPTS_ARG_ENUM_PURGE_STR,
+                     POMDP_SOLVE_OPTS_Enum_Purge_Str );
   PO_showUsageEnumType( file,
                      POMDP_SOLVE_OPTS_ARG_INC_PRUNE_STR,
                      POMDP_SOLVE_OPTS_Inc_Prune_Str );
@@ -371,16 +378,8 @@ POMDP_SOLVE_OPTS_showUsage( FILE* file, char* exec_name )
                      POMDP_SOLVE_OPTS_ARG_FG_SAVE_STR,
                      Boolean_Str );
   PO_showUsageEnumType( file,
-                     POMDP_SOLVE_OPTS_ARG_ENUM_PURGE_STR,
-                     POMDP_SOLVE_OPTS_Enum_Purge_Str );
-  PO_showUsageEnumType( file,
                      POMDP_SOLVE_OPTS_ARG_FG_TYPE_STR,
                      POMDP_SOLVE_OPTS_Fg_Type_Str );
-  fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_MCGS_TRAJ_ITER_COUNT_STR );
-  fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_MCGS_NUM_TRAJ_STR );
-  PO_showUsageEnumType( file,
-                     POMDP_SOLVE_OPTS_ARG_METHOD_STR,
-                     POMDP_SOLVE_OPTS_Method_Str );
   fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_FG_POINTS_STR );
   PO_showUsageEnumType( file,
                      POMDP_SOLVE_OPTS_ARG_FG_PURGE_STR,
@@ -388,46 +387,33 @@ POMDP_SOLVE_OPTS_showUsage( FILE* file, char* exec_name )
   PO_showUsageEnumType( file,
                      POMDP_SOLVE_OPTS_ARG_FG_NONNEG_REWARDS_STR,
                      Boolean_Str );
-  fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_MCGS_TRAJ_LENGTH_STR );
   fprintf( file, "\t%s <string>\n", POMDP_SOLVE_OPTS_ARG_GRID_FILENAME_STR );
-
-  /*******************************/
-  /* Optimization parameters  */
-  /*******************************/
-  fprintf( file, "Optimization options:\n" );
-
-  fprintf( file, "\t%s <double>\n", POMDP_SOLVE_OPTS_ARG_PRUNE_EPSILON_STR );
-  fprintf( file, "\t%s <double>\n", POMDP_SOLVE_OPTS_ARG_FG_EPSILON_STR );
-  fprintf( file, "\t%s <double>\n", POMDP_SOLVE_OPTS_ARG_LP_EPSILON_STR );
+  fprintf( file, "\t%s <string>\n", POMDP_SOLVE_OPTS_ARG_RAND_SEED_STR );
   PO_showUsageEnumType( file,
-                     POMDP_SOLVE_OPTS_ARG_DOM_CHECK_STR,
+                     POMDP_SOLVE_OPTS_ARG_FORCE_ROUNDING_STR,
                      Boolean_Str );
-  PO_showUsageEnumType( file,
-                     POMDP_SOLVE_OPTS_ARG_Q_PURGE_STR,
-                     POMDP_SOLVE_OPTS_Q_Purge_Str );
-  fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_ALG_RAND_STR );
-  fprintf( file, "\t%s <double>\n", POMDP_SOLVE_OPTS_ARG_EPSILON_STR );
-  PO_showUsageEnumType( file,
-                     POMDP_SOLVE_OPTS_ARG_PROJ_PURGE_STR,
-                     POMDP_SOLVE_OPTS_Proj_Purge_Str );
-  fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_PRUNE_RAND_STR );
-  PO_showUsageEnumType( file,
-                     POMDP_SOLVE_OPTS_ARG_WITNESS_POINTS_STR,
-                     Boolean_Str );
-
+  /* Not implemented yet!
+  fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_MCGS_TRAJ_ITER_COUNT_STR );
+  fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_MCGS_NUM_TRAJ_STR );
+  fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_MCGS_TRAJ_LENGTH_STR );
+  fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_MCGS_PRUNE_FREQ_STR );
+  */
+  
   /*******************************/
   /* Debug parameters  */
   /*******************************/
+  /* is now in general 
   fprintf( file, "Debug options:\n" );
 
   PO_showUsageEnumType( file,
                      POMDP_SOLVE_OPTS_ARG_VERBOSE_STR,
                      POMDP_SOLVE_OPTS_Verbose_Str );
+  */
 
   /*******************************/
   /* Value Iteration parameters  */
   /*******************************/
-  fprintf( file, "Value Iteration options:\n" );
+  fprintf( file, "\nValue Iteration options:\n" );
 
   fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_HISTORY_LENGTH_STR );
   PO_showUsageEnumType( file,
@@ -437,7 +423,6 @@ POMDP_SOLVE_OPTS_showUsage( FILE* file, char* exec_name )
   fprintf( file, "\t%s <double>\n", POMDP_SOLVE_OPTS_ARG_END_EPSILON_STR );
   fprintf( file, "\t%s <double>\n", POMDP_SOLVE_OPTS_ARG_START_EPSILON_STR );
   fprintf( file, "\t%s <double>\n", POMDP_SOLVE_OPTS_ARG_STOP_DELTA_STR );
-  fprintf( file, "\t%s <string>\n", POMDP_SOLVE_OPTS_ARG_POMDP_STR );
   PO_showUsageEnumType( file,
                      POMDP_SOLVE_OPTS_ARG_STOP_CRITERIA_STR,
                      POMDP_SOLVE_OPTS_Stop_Criteria_Str );
@@ -450,6 +435,31 @@ POMDP_SOLVE_OPTS_showUsage( FILE* file, char* exec_name )
                      POMDP_SOLVE_OPTS_Vi_Variation_Str );
   fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_HORIZON_STR );
   fprintf( file, "\t%s <double>\n", POMDP_SOLVE_OPTS_ARG_MAX_SOLN_SIZE_STR );
+
+  /*******************************/
+  /* Optimization parameters  */
+  /*******************************/
+  fprintf( file, "\nOptimization options:\n" );
+
+  fprintf( file, "\t%s <double>\n", POMDP_SOLVE_OPTS_ARG_EPSILON_STR );
+  fprintf( file, "\t%s <double>\n", POMDP_SOLVE_OPTS_ARG_PRUNE_EPSILON_STR );
+  fprintf( file, "\t%s <double>\n", POMDP_SOLVE_OPTS_ARG_FG_EPSILON_STR );
+  fprintf( file, "\t%s <double>\n", POMDP_SOLVE_OPTS_ARG_LP_EPSILON_STR );
+  PO_showUsageEnumType( file,
+                     POMDP_SOLVE_OPTS_ARG_DOM_CHECK_STR,
+                     Boolean_Str );
+  PO_showUsageEnumType( file,
+                     POMDP_SOLVE_OPTS_ARG_Q_PURGE_STR,
+                     POMDP_SOLVE_OPTS_Q_Purge_Str );
+  fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_ALG_RAND_STR );
+  PO_showUsageEnumType( file,
+                     POMDP_SOLVE_OPTS_ARG_PROJ_PURGE_STR,
+                     POMDP_SOLVE_OPTS_Proj_Purge_Str );
+  fprintf( file, "\t%s <int>\n", POMDP_SOLVE_OPTS_ARG_PRUNE_RAND_STR );
+  PO_showUsageEnumType( file,
+                     POMDP_SOLVE_OPTS_ARG_WITNESS_POINTS_STR,
+                     Boolean_Str );
+
 
 }  /* POMDP_SOLVE_OPTS_showUsage */